+++ tests/test_item.py
@@ -4,7 +4,7 @@ import pytest
 from pandas import Timestamp
 
 from OSmOSE.data.base_item import BaseItem
-from OSmOSE.utils.timestamp_utils import remove_overlaps
+from OSmOSE.utils.data_utils import remove_overlaps
 
 
 @pytest.mark.parametrize(
